# api/swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: 'WornOut'
  description: 'Catalog API that links items and users together'
host: virtserver.swaggerhub.com
basePath: /aeckard87/wornOut/1.0
...
schemes:
  - http
paths:
...
  /items:
    get:
      tags:
        - Items
      summary: Get all Items
      operationId: getItems
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Status Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/Item'
        '404':
          description: No items found
    delete:
      tags:
        - Items
      summary: Delete all Items
      operationId: deleteItems
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Status Ok
        '404':
          description: No items found
  /items/{id}:
    get:
      tags:
        - Items
      summary: Get specific Item
      operationId: getItem
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
          description: Item ID
      responses:
        '200':
          description: Status Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/Item'
        '404':
          description: No items found
    put:
      tags:
        - Items
      summary: update an Item
      operationId: updateItem
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
          description: Item ID
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Item'
      responses:
        '201':
          description: Category Created
          schema:
            $ref: '#/definitions/Item'
        '400':
          description: Invalid input
    delete:
      tags:
        - Items
      summary: Delete an Item
      operationId: deleteItem
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
          description: Item ID
      responses:
        '200':
          description: Status Ok
        '404':
          description: No items found
  '/users/{id}/items':
    get:
      tags:
        - Items
      summary: Get all Items by Owner
      operationId: getItemsByOwner
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
          description: Owner's ID
      responses:
        '200':
          description: Status Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/Item'
        '400':
          description: Invalid input
        '404':
          description: No items found
    post:
      tags:
        - Items
      summary: Create an Item
      operationId: createItem
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
          description: Owner's ID
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Item'
      responses:
        '201':
          description: Category Created
          schema:
            $ref: '#/definitions/Item'
        '400':
          description: Invalid input
    delete:
      tags:
        - Items
      summary: Delete an Item
      operationId: deleteItemsByOwner
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
          description: Item ID
      responses:
        '200':
          description: Status Ok
        '404':
          description: No items found
  '/subcategories/{id}/items':
    get:
      tags:
        - Items
      summary: Get all Items by Subcategory
      operationId: getItemsBySubCategory
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
          description: SubCategory ID
      responses:
        '200':
          description: Status Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/Item'
        '400':
          description: Invalid input
        '404':
          description: No items found
    delete:
      tags:
        - Items
      summary: Delete an Item
      operationId: deleteItemsBySubCategory
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
          description: SubCategory ID
      responses:
        '200':
          description: Status Ok
        '404':
          description: No items found
definitions:
...
  Descriptor:
    type: object
    properties:
      id:
        type: integer
        example: 89
      descriptor:
        type: string
        example: Ralph Lauren
      detailId:
        type: integer
        example: 10
...
  Item:
    type: object
    properties:
      id:
        type: integer
        example: 10
      name:
        type: string
        example: Red Short Sleeve Top
      userID:
        type: integer
        example: 40
      subCategoryID:
        type: integer
        example: 40
      descriptions:
        type: array
        items:
          $ref: '#/definitions/Descriptor'